+++ test/cppsuite/test_harness/util/scoped_types.h
